+++ routes/bot.py
@@ -1,14 +1,10 @@
-from framework import private, text
+from framework import private
 from framework.route import Route
 
 from app.Handlers.StartHandler import StartHandler
-from app.Handlers.SearchHandler import SearchHandler
 from app.Handlers.MessageHandler import MessageHandler
 
 
-Route.inline(SearchHandler.inline)
-Route.current(SearchHandler.current)
-
 # commands
 Route.command("start", StartHandler.start, private)
-Route.message(MessageHandler.message)
+Route.message(MessageHandler.message, private)